pyproject.toml
[project]
name = "hammy"
version = "0.0.1"
description = "Upload images to fapping"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["images", "upload", "fapping"]

authors = [
  {name = "brokedarius", email = "-@-.org" }
]

maintainers = [
  {name = "brokedarius", email = "-@-.org" }
]

classifiers = [
  "Development Status :: 5 - Production/Stable",

  "Intended Audience :: Developers",
  
  "Topic :: Software Development :: Build Tools",

  "License :: OSI Approved :: MIT License",

  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3 :: Only",
]

dependencies = [
  "msgspec",
  "pillow",
  "platformdirs",
  "pyperclip",
  "requests",
  "rich",
  "tomlkit",
  "urllib3",
  "pyperclip"
]

[project.scripts]
hammy = "hammy.__main__:main"

[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
"hammy.resize" = ["*.placeholder"]